Not-So-Almost There, Take 1...
So because I cannot let myself test only one thing, on top of trying to
solve for a "navbar" instead of "nav" element so that it locks to the
top, now I have added the complexity of a dropdown, which uses
javascript.  And that didn't take with the last commit.

So my theory now is that both problems may be caused by the default
Bootstrap being downloaded may not be the full one with all Javascript
plugins.  There is definitely a different filesize when I instead go to
Bootstrap>Customize>Download instead of DLing from Bootstrap's main
page.  Let's see if this solves anything...