Moving to Advanced Stuff
Adding the nav bars and pages for the advanced Bootstrap page and the
jQuery page.

Also fixing a bug on the homepage where jQuery wasn't called on the page
so the dropdown didn't work.  Goes to show that javascript alone in
Bootstrap does not work.  That explains the issues the team was working
to fix that I didn't understand yet.