Can Nav Be at the Bottom?
Moved the nav to the bottom of the TB Advanced page. Let's see if it
still hugs the top.  This structure would work better for plain text
pages, mobile sites, and anything that does not render CSS properly
(leaves the UL uptop).

Also fixed a bug with a link that never ended on the new pages.