Putting on the Polish - Personal Time
Still at work.  Decided to polish and close this project.  Since I built
the metaphorical cabinet on my own time, but based on the timber milled
earlier in the day, it is only right that I sweep up and laquer the
beauty in case anyone comes looking.  Like the trim?  Ok, this metaphor
is getting weak :P